refactor: add publish button component
