+++ src/app/dashboard/create/components/frame-preview.tsx
@@ -1,96 +1,22 @@
 import Image from "next/image";
-import { toast } from "sonner";
 import { useRef, useState } from "react";
-import { useMutation } from "@urql/next";
-import { PackageCheck } from "lucide-react";
-import { useRouter } from "next/navigation";
 import {
   ReactZoomPanPinchRef,
   TransformComponent,
   TransformWrapper,
 } from "react-zoom-pan-pinch";
 
-import { graphql } from "@/graphql";
 import { Card } from "@/components/ui/card";
 import { Label } from "@/components/ui/label";
 import { Input } from "@/components/ui/input";
-import { Button } from "@/components/ui/button";
-import { uploadFiles } from "@/lib/uploadthing";
 import { Textarea } from "@/components/ui/textarea";
 import { useFrameStore } from "@/store/useFrameStore";
-
-const CreateFrameMutation = graphql(`
-  mutation CreateFrame($input: CreateFrameInput!) {
-    createFrame(input: $input) {
-      id
-      title
-      caption
-      handle
-      imgUrl
-    }
-  }
-`);
+import { PublishButton } from "./publish-button";
 
 export function FramePreview() {
-  const router = useRouter();
-  const [{ fetching }, createFrameFn] = useMutation(CreateFrameMutation);
   const [frameOpacity, setFrameOpacity] = useState(1);
   const controlRef = useRef<ReactZoomPanPinchRef>(null);
   const frameData = useFrameStore((state) => state.frameData);
-
-  const isPublishing = useFrameStore((state) => state.isPublishing);
-  const updateIsPublishing = useFrameStore((state) => state.updateIsPublishing);
-
-  const handlePublish = () => {
-    if (!frameData.file) {
-      return;
-    }
-
-    updateIsPublishing(true);
-
-    toast.promise(
-      uploadFiles("imageUploader", {
-        files: [frameData.file],
-      }),
-      {
-        loading: "Please wait...",
-        success: (data) => {
-          toast.promise(
-            createFrameFn({
-              input: {
-                title: frameData.title,
-                handle: frameData.urlHandle,
-                caption: frameData.caption,
-                imgUrl: data[0].url,
-              },
-            }),
-            {
-              loading: "Publishing...",
-              success: (res) => {
-                if (res.error) {
-                  updateIsPublishing(false);
-                  return `Error publishing frame: ${res.error.message}`;
-                }
-
-                updateIsPublishing(false);
-                router.push("/dashboard");
-                return "Frame published";
-              },
-              error: (err) => {
-                updateIsPublishing(false);
-                return `Error publishing frame: ${err.message}`;
-              },
-            },
-          );
-          return "Almost there...";
-        },
-        error: (err) => {
-          updateIsPublishing(false);
-          return `Error uploading frame: ${err.message}`;
-        },
-      },
-    );
-  };
 
   return (
     <section className="w-full space-y-6">
@@ -162,14 +88,7 @@ export function FramePreview() {
           Confirm your frame details before publishing.
         </p>
 
-        <Button
-          disabled={fetching || isPublishing}
-          onClick={handlePublish}
-          className="w-full mt-4"
-        >
-          <PackageCheck className="mr-2 h-4 w-4" />
-          Publish Frame
-        </Button>
+        <PublishButton />
       </div>
     </section>
   );
